requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas>=2.1.4
requests==2.31.0
biopython>=1.85
openpyxl==3.1.2
selenium==4.15.2
pathlib2==2.3.7
numpy>=1.26.0
cgecore==2.0.1 